backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
openpyxl==3.1.2
joblib==1.3.2
python-whois==0.8.0
gunicorn==21.2.0
